// src/api/modrinth/validation.ts
import z from 'zod';

export const GetProject = z.object({
  slug: z.string(),
  title: z.string(),
  description: z.string(),
  categories: z.array(z.string()),
  client_side: z.enum(['required', 'optional', 'unsupported', 'unknown']),
  server_side: z.enum(['required', 'optional', 'unsupported', 'unknown']),
  body: z.string(),
  status: z.enum([
    'approved',
    'archived',
    'rejected',
    'draft',
    'unlisted',
    'processing',
    'withheld',
    'scheduled',
    'private',
    'unknown',
  ]),
  requested_status: z.enum(['approved', 'archived', 'unlisted', 'private', 'draft']).nullable(),
  additional_categories: z.array(z.string()),
  issues_url: z.string().nullable(),
  source_url: z.string().nullable(),
  wiki_url: z.string().nullable(),
  discord_url: z.string().nullable(),
  donation_urls: z.array(
    z.object({
      id: z.string(),
      platform: z.string(),
      url: z.string(),
    }),
  ),
  project_type: z.enum(['mod', 'modpack', 'resourcepack', 'shader']),
  downloads: z.int(),
  icon_url: z.string().nullable(),
  color: z.int().nullable(),
  thread_id: z.string(),
  monetization_status: z.enum(['monetized', 'demonetized', 'force-demonitized']),
  id: z.string(),
  team: z.string(),
  body_url: z.string().nullable(),
  moderator_message: z.unknown().nullable(),
  published: z.iso.datetime(),
  updated: z.iso.datetime(),
  approved: z.iso.datetime().nullable(),
  queued: z.iso.datetime().nullable(),
  followers: z.int(),
  license: z.object({
    id: z.string(),
    name: z.string(),
    url: z.string(),
  }),
  versions: z.array(z.string()),
  game_versions: z.array(z.string()),
  loaders: z.array(z.string()),
  gallery: z.array(z.unknown()),
});

export type GetProjectT = z.infer<typeof GetProject>;